change drupal_bootstrap() to use a Pimple container.
drupal_bootstrap() sets up the Drupal run time environment with all of
its parameters and services. This function is response for PHP and
environment settings, sessions, locale, database, page cache, page
headers, Drupal site variables and more.

drupal_bootstrap() wants to be failsafe, but as a result of all these
responsibilities, it is very complicated and full of interdependencies.
Changing this function to use a Pimple container will let the bootstrap
process be extended, overridden and rewritten.